Moves the file in parsing packages for better clarity
