+++ app/src/main/java/net/squanchy/tweets/parsing/ParsingRegex.java
@@ -1,9 +1,9 @@
-package net.squanchy.tweets.util;
+package net.squanchy.tweets.parsing;
 
 import android.support.annotation.StringDef;
 
-import static net.squanchy.tweets.util.ParsingRegex.HASHTAG_REGEX;
-import static net.squanchy.tweets.util.ParsingRegex.MENTIONS_REGEX;
+import static net.squanchy.tweets.parsing.ParsingRegex.HASHTAG_REGEX;
+import static net.squanchy.tweets.parsing.ParsingRegex.MENTIONS_REGEX;
 
 @StringDef({MENTIONS_REGEX, HASHTAG_REGEX})
 public @interface ParsingRegex {
