+++ app/src/test/java/net/squanchy/tweets/parsing/SpannableDataExtractorTest.java
@@ -1,10 +1,9 @@
-package net.squanchy.tweets.util;
+package net.squanchy.tweets.parsing;
 
 import net.squanchy.tweets.model.TweetSpecialTextData;
 
 import org.junit.Test;
 
-import static net.squanchy.tweets.util.TweetTestData.MentionsAndUrls.TEST_TWEET_WITh_MENTIONS_AND_URLS;
 import static org.junit.Assert.assertEquals;
 
 
@@ -16,7 +15,7 @@ public class SpannableDataExtractorTest {
         int end = 34;
         String text = "#droidconit";
 
-        TweetSpecialTextData data = SpannableDataExtractor.extract(TEST_TWEET_WITh_MENTIONS_AND_URLS, start, end);
+        TweetSpecialTextData data = SpannableDataExtractor.extract(TweetTestData.MentionsAndUrls.TEST_TWEET_WITh_MENTIONS_AND_URLS, start, end);
 
         assertEquals(text, data.text());
     }
